Add TimeWindow transform
